+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support for `dns_tsig_key` resource and datasources
 - Support for Economy vaults in Key management service
 - documentation update for Oracle Kubernetes engine
+- Support for API Gateway Service
 
 ### Notes
 Starting with this version, the terraform-provider-oci supports VM Instance resizing with reboot. Resizing can only happen within the shapes of same family. The shapes much be compatible with the image and the instance should not be associated to any `dedicated_vm_host_id`.